# fix keyword-prefixed name parsing