Fixes https://github.com/flix/flix/issues/929

The parser greedily accepts `true`/`false`/`null` as an expression and doesn't know what to do when it turns out it wasn't a keyword after all.

I added a rule `keyword` to the parser, which ensures the keyword isn't just a prefix of a longer name. I only applied it to a couple places where I saw it was needed, but if we like this solution, it might save us some trouble in the future to use it for all keywords.