@magnus-madsen 2020-08-26
I like this solution. I would like to apply it everywhere we have an `atomic`.

Does it impact parsing performance measurably?