@mlutze 2020-08-26
> I would like to apply it everywhere we have an `atomic`.

I don't think we should replace `atomic` entirely, since this would disallow cases like `x<=y` (no spaces). Maybe not the best formatting, but should still parse.

> Does it impact parsing performance measurably?

I haven't been able to get consistent results with the performance tests, so it's hard to be certain. I've dumped results of three (per branch) test runs below.

<details>
<summary>Test 1</summary>

master throughput | feature throughput | absolute difference | percent difference
--- | --- | --- | ---
18821 | 18365 | -456 | -2.4%

phase | master time | feature time | absolute difference | percent difference
--- | --- | --- | --- | ---
Reader | 845240 | 900074 | +54834 | +6.5%
Parser | 136379270 | 172842006 | +36462736 | +26.7%
Weeder | 13360179 | 13872081 | +511902 | +3.8%
Namer | 44953838 | 38331724 | -6622114 | -14.7%
Resolver | 66090897 | 70139305 | +4048408 | +6.1%
Typer | 292516142 | 334528895 | +42012753 | +14.4%
Stratifier | 12347328 | 19553973 | +7206645 | +58.4%
PatternExhaustiveness | 31042931 | 35609338 | +4566407 | +14.7%
Redundancy | 23152975 | 26660376 | +3507401 | +15.1%
Safety | 4296650 | 4335169 | +38519 | +0.9%
Documentor | 2322 | 2053 | -269 | -11.6%
Monomorph | 279988024 | 263685082 | -16302942 | -5.8%
Synthesize | 12484702 | 12944861 | +460159 | +3.7%
Simplifier | 78964685 | 79206568 | +241883 | +0.3%
ClosureConv | 26264890 | 28109206 | +1844316 | +7.0%
LambdaLift | 10259237 | 9413658 | -845579 | -8.2%
Tailrec | 3980261 | 4164829 | +184568 | +4.6%
Optimizer | 11006252 | 11113203 | +106951 | +1.0%
TreeShaker | 41404275 | 44582009 | +3177734 | +7.7%
VarNumbering | 7189058 | 7173707 | -15351 | -0.2%
Finalize | 30092216 | 32309959 | +2217743 | +7.4%
JvmBackend | 702729283 | 705969973 | +3240690 | +0.5%
</details>

<details>
<summary>Test 2</summary>

master throughput | feature throughput | absolute difference | percent difference
--- | --- | --- | ---
18043 | 19078 | +1035 | +5.7%

phase | master time | feature time | absolute difference | percent difference
--- | --- | --- | --- | ---
Reader | 878520 | 1074569 | +196049 | +22.3%
Parser | 126104460 | 237853911 | +111749451 | +88.6%
Weeder | 15351026 | 16650417 | +1299391 | +8.5%
Namer | 41636608 | 43871497 | +2234889 | +5.4%
Resolver | 90206140 | 94760536 | +4554396 | +5.0%
Typer | 289974397 | 423845564 | +133871167 | +46.2%
Stratifier | 13373883 | 13443253 | +69370 | +0.5%
PatternExhaustiveness | 30861451 | 32820913 | +1959462 | +6.3%
Redundancy | 28171566 | 38461706 | +10290140 | +36.5%
Safety | 4852707 | 6175592 | +1322885 | +27.3%
Documentor | 2461 | 2273 | -188 | -7.6%
Monomorph | 261405258 | 324881503 | +63476245 | +24.3%
Synthesize | 12720261 | 16202917 | +3482656 | +27.4%
Simplifier | 81254661 | 100997303 | +19742642 | +24.3%
ClosureConv | 29164676 | 33239713 | +4075037 | +14.0%
LambdaLift | 10449689 | 11937809 | +1488120 | +14.2%
Tailrec | 4085961 | 5107627 | +1021666 | +25.0%
Optimizer | 10978990 | 13347699 | +2368709 | +21.6%
TreeShaker | 46228318 | 57330085 | +11101767 | +24.0%
VarNumbering | 7242331 | 8698348 | +1456017 | +20.1%
Finalize | 32208749 | 40019360 | +7810611 | +24.2%
JvmBackend | 687686760 | 1062486131 | +374799371 | +54.5%
</details>

<details>
<summary>Test 3</summary>

master throughput | feature throughput | absolute difference | percent difference
--- | --- | --- | ---
19162 | 18975 | -187 | -1.0%

phase | master time | feature time | absolute difference | percent difference
--- | --- | --- | --- | ---
Reader | 934989 | 926177 | -8812 | -0.9%
Parser | 162965275 | 109719663 | -53245612 | -32.7%
Weeder | 11832485 | 15026977 | +3194492 | +27.0%
Namer | 38633260 | 39552495 | +919235 | +2.4%
Resolver | 71247645 | 83816766 | +12569121 | +17.6%
Typer | 356848250 | 319990987 | -36857263 | -10.3%
Stratifier | 12394340 | 12976133 | +581793 | +4.7%
PatternExhaustiveness | 30048456 | 33622063 | +3573607 | +11.9%
Redundancy | 24356878 | 30734129 | +6377251 | +26.2%
Safety | 4575963 | 4559172 | -16791 | -0.4%
Documentor | 2287 | 2457 | +170 | +7.4%
Monomorph | 289801986 | 261063774 | -28738212 | -9.9%
Synthesize | 12639434 | 13909249 | +1269815 | +10.0%
Simplifier | 104349874 | 81587265 | -22762609 | -21.8%
ClosureConv | 28340802 | 24846589 | -3494213 | -12.3%
LambdaLift | 8645399 | 9479387 | +833988 | +9.6%
Tailrec | 4052432 | 4069430 | +16998 | +0.4%
Optimizer | 12383538 | 11175340 | -1208198 | -9.8%
TreeShaker | 43172758 | 43571735 | +398977 | +0.9%
VarNumbering | 6829366 | 7109941 | +280575 | +4.1%
Finalize | 27083766 | 33607329 | +6523563 | +24.1%
JvmBackend | 707995178 | 781880701 | +73885523 | +10.4%
</details>