@magnus-madsen 2020-08-27
OK; maybe I was overeager. When exactly should we be using `keyword`?