@mlutze 2020-08-27
I think using it only for full-word, alphabetic tokens is the best strategy, since that's the only place a keyword could be a prefix of a name. (right?)

For example:
* `atomic("<=")` because `<=` cannot be the prefix of a name
* `keyword("let")` because `let` could be the prefix of name `lettuce`
* `atomic("@")` because `@` must be the prefix of some `@annotation`

An interesting case is `let*`. Do we want to parse for example `let*x = f(y)`? Or should we require the space? (I would guess the parser already says there has to be whitespace after `let*` but the principle remains in question.)