@magnus-madsen 2020-08-27
> An interesting case is `let*`. Do we want to parse for example `let*x = f(y)`? Or should we require the space? (I would guess the parser already says there has to be whitespace after `let*` but the principle remains in question.)

This is what confuses me a bit. In many (most? all?) places we already demand hard whitespace (WS) to ensure that the keyword does not run into some other text.