@mlutze 2020-08-27
It seems to me that the reason we had issues with `null/false/true` is that in the parser, they were individual rules, so no whitespace match could be put after them (because we want to be able to match e.g. `{a = true}`). In contrast, `let` only comes as part of a larger rule so it always has the `WS` after it.

So it's not really necessary to put `keyword` on all keywords, but doing so might make us more robust to parser changes in the future.